Improvements for the OSM importer. Taking into account more tags especially related to user class accessibility (access, foot, bicycle, motor_vehicles...) leading to cleverer road journey planning in general.
Improving greatly definition of a central road chunk for a city. Update at each OSM import and don't search to create one if there is a main stop area in the city. As for physical stop projection, ensure that the road chunk chosen is accessible to every user class.

Please reimport OSM datas after this commit to benefit from all the improvements.

+++ src/06_openstreetmap/OSMNetwork.cpp
@@ -170,13 +170,13 @@ std::map<unsigned long long int,std::pair<RelationPtr,std::map<unsigned long lon
 
    util::Log::GetInstance().info("extracting ways by administrative boundaries");
    /*
-   std::map<int,RelationPtr>::iterator boundaryIterator = adminBoundaries.begin();
+   std::map<unsigned long long int,RelationPtr>::iterator boundaryIterator = adminBoundaries.begin();
    while(boundaryIterator != adminBoundaries.end()) {
       RelationPtr boundary = boundaryIterator->second;
-      ret[boundaryIterator->first] = std::pair<RelationPtr,std::map<int,WayPtr> >(boundary, std::map<int,WayPtr>());
+      ret[boundaryIterator->first] = std::pair<RelationPtr,std::map<unsigned long long int,WayPtr> >(boundary, std::map<unsigned long long int,WayPtr>());
       boost::shared_ptr<const geos::geom::prep::PreparedGeometry> boundaryPrepGeom =
             boundary->toPreparedGeometry();
-      std::map<int,WayPtr>::iterator wayIterator = walkableWays.begin();
+      std::map<unsigned long long int,WayPtr>::iterator wayIterator = walkableWays.begin();
       while(wayIterator != walkableWays.end()) {
          WayPtr way = wayIterator->second;
          boost::shared_ptr<const geos::geom::Geometry> wayGeom = way->toGeometry();
